requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
httpx
pillow
python-multipart
pydantic-settings
groq
sentence-transformers